A modified patch from  Anahita Goljahani  for "Overview of Bills of Material doesn't paginate" https://issues.apache.org/jira/browse/OFBIZ-5324
When having more than 20 BoM's in Manufacturing the overview doesn't paginate.

jleroux; I'm not quite sure why an orderBy was necessary but it works and is a simple enough solution. I refactored things a bit and kept find instead of proposed findList. So this just add an orderBy to the find

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/manufacturing@1524676 13f79535-47bb-0310-9956-ffa450edef68