A patch from Pierre Smits for <<Correct setting of the WorkEffortPurposeTypeId on production run tasks>> https://issues.apache.org/jira/browse/OFBIZ-6954
Currently when a new production run is created, the workeffort purpose type id is set to WEPT_PRODUCTION_RUN. This setting should also be applied to the associate production run tasks. 
However, this doesn't happen. Instead the workeffort purpose type id from the routing task is copied.

jleroux: I agree with Pierre after a short discussion. This patch will ensure that routing tasks will be visible in the CAL_MANUFACTURING calendar without losing any information since the association, be a routing task or a task for a production run, will remain in the WorkEffortAssoc entity with a workEffortAssocTypeId = “WORK_EFF_TEMPLATE”.




git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/manufacturing@1736621 13f79535-47bb-0310-9956-ffa450edef68