Fixes a bug introduced with r1737531 for OFBIZ-6817 (manufacturing). Part of "Move all Groovy scripts in components WEB-INF/actions sub-folders" - https://issues.apache.org/jira/browse/OFBIZ-6981
Currently there are few Groovy scripts not in components WEB-INF/actions sub-folders.

Before moving Groovy scripts as we decided in OFBIZ-6810 we should move the few Groovy scripts not in components WEB-INF/actions sub-folders to these WEB-INF/actions sub-folders. So we will not miss one and it will be a "straightforward" task to move then as we decided in OFBIZ-6810.

The 1st task here will be to fix a bug related with ShowProductionRun.groovy. 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/applications/manufacturing@1737603 13f79535-47bb-0310-9956-ffa450edef68